"Add @radix-ui/react-avatar and update SideMenu component to include Post component in resizable panel"
This commit adds the package `@radix-ui/react-avatar` and its dependencies to the project. The `SideMenu` component has been updated to include a new `Post` component in the resizable panel when the panel size is greater than 9. The `TooltipProvider` and `Tooltip` components from `@radix-ui/react-tooltip` have also been added to the project and used in the `SideMenu` component to display tooltips for the side menu items when the panel size is less than or equal to 9. The `SignOut` icon is now only displayed in the button when the panel size is greater than 9.